rewrite xfstest

+++ fs/xfstests.py
@@ -28,7 +28,7 @@ import shutil
 
 from avocado import Test
 from avocado.utils import process, build, git, distro, partition
-from avocado.utils import disk, data_structures, pmem
+from avocado.utils import disk, pmem
 from avocado.utils import genio
 from avocado.utils.software_manager.manager import SoftwareManager
 
@@ -61,7 +61,6 @@ class Xfstests(Test):
         return namespace_size
 
     def setup_nvdimm(self):
-        self.logflag = self.params.get('logdev', default=False)
         self.plib = pmem.PMem()
         self.plib.enable_region()
         regions = sorted(self.plib.run_ndctl_list('-R'),
@@ -149,18 +148,7 @@ class Xfstests(Test):
             self.scratch_dev = "/dev/%s" % pmem_dev
             self.devices.extend([self.test_dev, self.scratch_dev])
 
-    def setUp(self):
-        """
-        Build xfstest
-        Source: git://git.kernel.org/pub/scm/fs/xfs/xfstests-dev.git
-        """
-        self.use_dd = False
-        root_fs = process.system_output(
-            "df -T / | awk 'END {print $2}'", shell=True).decode("utf-8")
-        if root_fs in ['ext3', 'ext4']:
-            self.use_dd = True
-        self.dev_type = self.params.get('type', default='loop')
-
+    def __setUp_packages(self):
         sm = SoftwareManager()
 
         self.detected_distro = distro.detect()
@@ -187,7 +175,7 @@ class Xfstests(Test):
                 packages.extend(['ndctl', 'parted'])
                 if self.detected_distro.name == 'rhel':
                     packages.extend(['daxctl'])
-            packages.extend(['acl', 'bc', 'dump', 'indent', 'libtool', 'lvm2',
+            packages.extend(['acl', 'bc', 'indent', 'libtool', 'lvm2',
                              'xfsdump', 'psmisc', 'sed', 'libacl-devel',
                              'libattr-devel', 'libaio-devel', 'libuuid-devel',
                              'openssl-devel', 'xfsprogs-devel', 'gettext',
@@ -208,7 +196,6 @@ class Xfstests(Test):
                 packages = list(set(packages)-set(packages_remove))
             elif self.detected_distro.name == 'rhel' and\
                     self.detected_distro.version.startswith('9'):
-                packages_remove.extend(['dump'])
                 packages = list(set(packages)-set(packages_remove))
 
             if self.detected_distro.name in ['centos', 'fedora']:
@@ -220,22 +207,42 @@ class Xfstests(Test):
             if not sm.check_installed(package) and not sm.install(package):
                 self.cancel("Fail to install %s required for this test." %
                             package)
+
+    def setUp(self):
+        """
+        Build xfstest
+        Source: git://git.kernel.org/pub/scm/fs/xfs/xfstests-dev.git
+        """
+        self.use_dd = False
+        root_fs = process.system_output(
+            "df -T / | awk 'END {print $2}'", shell=True).decode("utf-8")
+
+        if root_fs in ['ext2', 'ext3']:
+            self.use_dd = True
+
+        self.dev_type = self.params.get('type', default='loop')
+
+        self.__setUp_packages()
+
+        self.fs_to_test = self.params.get('fs', default='ext4')
+        self.args = self.params.get('args', default='-g auto')
+        self.sectionname = self.params.get('sectionname', default=self.fs_to_test)
         self.skip_dangerous = self.params.get('skip_dangerous', default=True)
-        self.group = self.params.get('group', default='auto')
-        self.test_range = self.params.get('test_range', default=None)
         self.base_disk = self.params.get('disk', default=None)
         self.scratch_mnt = self.params.get(
             'scratch_mnt', default='/mnt/scratch')
         self.test_mnt = self.params.get('test_mnt', default='/mnt/test')
         self.disk_mnt = self.params.get('disk_mnt', default='/mnt/loop_device')
-        self.fs_to_test = self.params.get('fs', default='ext4')
         self.run_type = self.params.get('run_type', default='distro')
+        self.mkfs_opt = self.params.get('mkfs_opt', default='')
+        self.mount_opt = self.params.get('mount_opt', default='')
+        self.logflag = self.params.get('logdev', default=False)
+        self.log_test = self.params.get('log_test', default='')
+        self.log_scratch = self.params.get('log_scratch', default='')
+        self.logdev_opt = self.params.get('logdev_opt', default='')
 
         self.devices = []
         self.part = None
-        if self.group and self.test_range:
-            self.cancel("incorrect yaml parameter, group and test range can"
-                        "not be run at same time")
 
         if self.run_type == 'upstream':
             prefix = "/usr/local"
@@ -323,19 +330,22 @@ class Xfstests(Test):
             self.cancel('Unknown filesystem %s' % self.fs_to_test)
         mount = True
         self.log_devices = []
+
+        # number of loops to create based on fs to test
+        self.num_loops = 2
+        if self.fs_to_test == "btrfs":
+            self.num_loops = 5
+
         shutil.copyfile(self.get_data('local.config'),
                         os.path.join(self.teststmpdir, 'local.config'))
         shutil.copyfile(self.get_data('group'),
                         os.path.join(self.teststmpdir, 'group'))
 
-        self.log_test = self.params.get('log_test', default='')
-        self.log_scratch = self.params.get('log_scratch', default='')
-
         if self.dev_type == 'loop':
             loop_size = self.params.get('loop_size', default='7GiB')
             if not self.base_disk:
                 # Using root for file creation by default
-                check = (int(loop_size.split('GiB')[0]) * 2) + 1
+                check = (int(loop_size.split('GiB')[0]) * self.num_loops) + 1
                 if disk.freespace('/') / 1073741824 > check:
                     self.disk_mnt = ''
                     mount = False
@@ -351,8 +361,6 @@ class Xfstests(Test):
         # mkfs for devices
         if self.devices:
             cfg_file = os.path.join(self.teststmpdir, 'local.config')
-            self.mkfs_opt = self.params.get('mkfs_opt', default='')
-            self.mount_opt = self.params.get('mount_opt', default='')
             with open(cfg_file, "r") as sources:
                 lines = sources.readlines()
             with open(cfg_file, "w") as sources:
@@ -366,9 +374,16 @@ class Xfstests(Test):
                             re.sub(r'export TEST_DIR=.*', 'export TEST_DIR=%s'
                                    % self.test_mnt, line))
                     elif line.startswith('export SCRATCH_DEV'):
-                        sources.write(re.sub(
-                            r'export SCRATCH_DEV=.*', 'export SCRATCH_DEV=%s'
-                                                      % self.devices[1], line))
+                        if self.fs_to_test == "btrfs":
+                            scratch_dev_pool = ' '.join(
+                                [(self.devices[i]) for i in range(1, 5)])
+                            sources.write(re.sub(r'export SCRATCH_DEV=.*',
+                                                 'export SCRATCH_DEV_POOL="%s"'
+                                                 % scratch_dev_pool, line))
+                        else:
+                            sources.write(re.sub(
+                                r'export SCRATCH_DEV=.*', 'export SCRATCH_DEV=%s'
+                                % self.devices[1], line))
                     elif line.startswith('export SCRATCH_MNT'):
                         sources.write(
                             re.sub(
@@ -390,7 +405,6 @@ class Xfstests(Test):
                     sources.write('MKFS_OPTIONS="%s"\n' % self.mkfs_opt)
                 if self.mount_opt:
                     sources.write('MOUNT_OPTIONS="%s"\n' % self.mount_opt)
-            self.logdev_opt = self.params.get('logdev_opt', default='')
             for dev in self.log_devices:
                 dev_obj = partition.Partition(dev)
                 dev_obj.mkfs(fstype=self.fs_to_test, args=self.mkfs_opt)
@@ -405,27 +419,12 @@ class Xfstests(Test):
         git.get_repo('git://git.kernel.org/pub/scm/fs/xfs/xfstests-dev.git',
                      destination_dir=self.teststmpdir)
 
-        build.make(self.teststmpdir)
+        extra_args = f"-j{os.cpu_count()}"
+        build.make(self.teststmpdir, extra_args=extra_args)
         self.available_tests = self._get_available_tests()
-
-        self.test_list = self._create_test_list(self.test_range)
         self.log.info("Tests available in srcdir: %s",
                       ", ".join(self.available_tests))
-        if not self.test_range:
-            self.exclude = self.params.get('exclude', default=None)
-            self.gen_exclude = self.params.get('gen_exclude', default=None)
-            self.share_exclude = self.params.get('share_exclude', default=None)
-            if self.exclude or self.gen_exclude or self.share_exclude:
-                self.exclude_file = os.path.join(self.teststmpdir, 'exclude')
-                if self.exclude:
-                    self._create_test_list(self.exclude, self.fs_to_test,
-                                           dangerous=False)
-                if self.gen_exclude:
-                    self._create_test_list(self.gen_exclude, "generic",
-                                           dangerous=False)
-                if self.share_exclude:
-                    self._create_test_list(self.share_exclude, "shared",
-                                           dangerous=False)
+
         if self.detected_distro.name is not 'SuSE':
             if process.system('useradd 123456-fsgqa', sudo=True, ignore_status=True):
                 self.log.warn('useradd 123456-fsgqa failed')
@@ -444,36 +443,27 @@ class Xfstests(Test):
     def test(self):
         failures = False
         os.chdir(self.teststmpdir)
-        if not self.test_list:
-            self.log.info('Running all tests')
-            args = ''
-            if self.exclude or self.gen_exclude:
-                args = ' -E %s' % self.exclude_file
-            cmd = './check %s -g %s' % (args, self.group)
-            result = process.run(cmd, ignore_status=True, verbose=True)
-            if result.exit_status == 0:
-                self.log.info('OK: All Tests passed.')
-            else:
-                msg = self._parse_error_message(result.stdout)
-                self.log.info('ERR: Test(s) failed. Message: %s', msg)
-                failures = True
-
+        cmd = './check %s' % self.args
+        result = process.run(cmd, ignore_status=True, verbose=True)
+        if result.exit_status == 0:
+            self.log.info("OK: All tests passed")
         else:
-            self.log.info('Running only specified tests')
-            for test in self.test_list:
-                test = '%s/%s' % (self.fs_to_test, test)
-                cmd = './check %s' % test
-                result = process.run(cmd, ignore_status=True, verbose=True)
-                if result.exit_status == 0:
-                    self.log.info('OK: Test %s passed.', test)
-                else:
-                    msg = self._parse_error_message(result.stdout)
-                    self.log.info('ERR: %s failed. Message: %s', test, msg)
-                    failures = True
+            msg = self._parse_error_message(result.stdout)
+            failures = True
+
         if failures:
-            self.fail('One or more tests failed. Please check the logs.')
+            self.fail("Test(s) failed %s" % msg)
 
     def tearDown(self):
+        srcdir = f"{self.teststmpdir}/results"
+        outputpath = f"{self.outputdir}/results-{self.sectionname}"
+        self.log.debug("srcdir: %s, outputpath: %s" % (srcdir, outputpath))
+        if (os.path.exists(self.outputdir)):
+            shutil.copytree(srcdir, outputpath)
+        else:
+            self.log.info("Unable to copy. Path not found %s -> %s " %
+                          (srcdir, self.outputdir))
+
         user_exits = 0
         if not (process.system('id fsgqa', sudo=True, ignore_status=True)):
             process.system('userdel -r -f fsgqa', sudo=True)
@@ -491,6 +481,8 @@ class Xfstests(Test):
             shutil.rmtree(self.scratch_mnt)
         if os.path.exists(self.test_mnt):
             shutil.rmtree(self.test_mnt)
+        if os.path.exists(self.disk_mnt):
+            shutil.rmtree(self.disk_mnt)
         if os.path.exists(self.teststmpdir + "/libini"):
             shutil.rmtree(self.teststmpdir + "/libini")
         if self.dev_type == 'loop':
@@ -512,49 +504,45 @@ class Xfstests(Test):
                             process.system('losetup -d %s' % dev, shell=True,
                                            sudo=True, ignore_status=True)
 
+    def _create_fsimages(self, loop_size, i):
+        dd_count = int(loop_size.split('GiB')[0])
+        fpath = f"{self.disk_mnt}/file-{i}.img"
+
+        # if file already present of the given size then just continue
+        # Note there is an inherent assumption anyways that the filesize
+        # is given in GiB
+        if (os.path.isfile(fpath) and
+                os.path.getsize(fpath) == (dd_count * 1024 * 1024 * 1024)):
+            self.log.debug("%s already present, continue" % (fpath))
+            return
+
+        if self.use_dd:
+            process.run('dd if=/dev/zero of=%s/file-%s.img bs=1G count=%s'
+                        % (self.disk_mnt, i, dd_count), shell=True,
+                        sudo=True)
+        else:
+            process.run('fallocate -o 0 -l %s %s/file-%s.img' %
+                        (loop_size, self.disk_mnt, i), shell=True,
+                        sudo=True)
+
     def _create_loop_device(self, loop_size, mount=True):
         if mount:
             self.part = partition.Partition(
                 self.base_disk, mountpoint=self.disk_mnt)
             self.part.mount()
-        # Creating two loop devices
-        for i in range(2):
-            if self.use_dd:
-                dd_count = int(loop_size.split('GiB')[0])
-                process.run('dd if=/dev/zero of=%s/file-%s.img bs=1G count=%s'
-                            % (self.disk_mnt, i, dd_count), shell=True,
-                            sudo=True)
-            else:
-                process.run('fallocate -o 0 -l %s %s/file-%s.img' %
-                            (loop_size, self.disk_mnt, i), shell=True,
-                            sudo=True)
+
+        # Creating loop devices
+        for i in range(self.num_loops):
+            self._create_fsimages(loop_size, i)
             dev = process.system_output('losetup -f').decode("utf-8").strip()
             self.devices.append(dev)
             process.run('losetup %s %s/file-%s.img' %
                         (dev, self.disk_mnt, i), shell=True, sudo=True)
 
-    def _create_test_list(self, test_range, test_type=None, dangerous=True):
-        test_list = []
+    def _create_test_list(self, dangerous=True):
         dangerous_tests = []
         if self.skip_dangerous:
             dangerous_tests = self._get_tests_for_group('dangerous')
-        if test_range:
-            for test in data_structures.comma_separated_ranges_to_list(test_range):
-                test = "%03d" % test
-                if dangerous:
-                    if test in dangerous_tests:
-                        self.log.debug('Test %s is dangerous. Skipping.', test)
-                        continue
-                if not self._is_test_valid(test):
-                    self.log.debug('Test %s invalid. Skipping.', test)
-                    continue
-                test_list.append(test)
-
-        if test_type:
-            with open(self.exclude_file, 'a') as fp:
-                for test in test_list:
-                    fp.write('%s/%s\n' % (test_type, test))
-        return test_list
 
     def _get_tests_for_group(self, group):
         """
@@ -579,8 +567,6 @@ class Xfstests(Test):
         os.chdir(self.teststmpdir)
         tests_set = []
         tests = glob.glob(self.teststmpdir + '/tests/*/???.out')
-        tests += glob.glob(self.teststmpdir + '/tests/*/???.out.linux')
-        tests = [t.replace('.linux', '') for t in tests]
 
         tests_set = sorted([t[-7:-4] for t in tests if os.path.exists(t[:-4])])
         tests_set = set(tests_set)
